:zap: use deck config

+++ scripts/cli.js
@@ -12,7 +12,8 @@ const cli = meow(
     $ ${chalk.green('next-mdx-deck deck.mdx')}
     $ ${chalk.green('next-mdx-deck build deck.mdx')}
   ${chalk.gray('Options')}
-      -p --port     Dev server port
+      -p --port     Dev server port (default: 8000)
+      -config       Relative location to config file (default: deck.config.js)
 `,
   {
     description: chalk.blue('next-mdx-deck'),
@@ -22,13 +23,18 @@ const cli = meow(
         alias: 'p',
         default: '8000',
       },
+      config: {
+        type: 'string',
+        alias: 'config',
+        default: 'deck.config.js',
+      },
     },
   }
 )
 /*
 {
     input: ['develop', 'build'],
-    flags: {port: '8000'},
+    flags: {port: '8000', config: 'deck.config.js' },
 }
 */
 
@@ -62,15 +68,32 @@ const copyAssetsFolder = async () => {
 const next = async (...args) => {
   // Get slides directory
   const slidesDir = path.join(__dirname, '../slides')
-  await fs.emptyDirSync(slidesDir)
 
   // Take MDX file from input and copy it into slides directory
   const mdxInputPath = `${process.cwd()}/${filename}`
   const newMdxFilePath = `${slidesDir}/${filename}`
+  const mdxFileExists = await fs.pathExists(newMdxFilePath)
 
-  await fs.copy(mdxInputPath, newMdxFilePath, (err) => {
-    if (err) return console.error(err)
-  })
+  if (mdxInputPath !== newMdxFilePath && mdxFileExists) {
+    // Delete old slides
+    await fs.emptyDirSync(slidesDir)
+    // Copy MDX file
+    await fs.copy(mdxInputPath, newMdxFilePath, (err) => {
+      if (err) return console.error(err)
+    })
+  }
+
+  const oldConfigPath = path.join(__dirname, '../deck.config.js')
+  const inputFilename = opts.config ? opts.config : 'deck.config.js'
+  const inputConfigPath = `${process.cwd()}/${inputFilename}`
+  const inputConfigExists = await fs.pathExists(inputConfigPath)
+  // Delete old config and copy
+  if (oldConfigPath !== inputConfigPath && inputConfigExists) {
+    await fs.remove(oldConfigPath)
+    await fs.copy(inputConfigPath, oldConfigPath, (err) => {
+      if (err) return console.error(err)
+    })
+  }
 
   // Copy contents of public folder (image assets and whatnot)
   // @TODO: Maybe pass Next CLI a flag for public dir?
@@ -95,6 +118,6 @@ switch (cmd) {
     break
   case 'dev':
   default:
-    next('dev', '-p', opts.port, opts.open && '--open')
+    next('dev', '-p', opts.port)
     break
 }
